# be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/event-groups": {
      "post": {
        "operationId": "EventGroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event group created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create event group",
        "tags": [
          "Event Groups"
        ]
      },
      "get": {
        "operationId": "EventGroupsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of event groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventGroupResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List event groups",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}": {
      "get": {
        "operationId": "EventGroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event group found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get event group by ID",
        "tags": [
          "Event Groups"
        ]
      },
      "patch": {
        "operationId": "EventGroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update event group",
        "tags": [
          "Event Groups"
        ]
      },
      "delete": {
        "operationId": "EventGroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event group deleted"
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete event group",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}/periods": {
      "post": {
        "operationId": "EventGroupsController_createPeriod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Period created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create period for event group",
        "tags": [
          "Event Groups"
        ]
      }
    },
    "/api/v1/event-groups/{id}/details": {
      "get": {
        "operationId": "EventGroupsController_getDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event group UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupDetailsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get event group details with analytics",
        "tags": [
          "Event Groups"
        ]
      }
    },
...
    "/api/v1/days/{date}/media": {
      "post": {
        "operationId": "MediaController_addMedia",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDayMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayMediaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attach uploaded media to a day",
        "tags": [
          "Media"
        ]
      },
      "get": {
        "operationId": "MediaController_getMediaForDay",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayMediaResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List media for a specific day",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/media/{id}": {
      "delete": {
        "operationId": "MediaController_deleteMedia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media item UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Media deleted"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a media item",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/memories/on-this-day": {
      "get": {
        "description": "Returns memories from 1 month, 6 months, and 1 year ago for the given date (defaults to today).",
        "operationId": "MemoriesController_getOnThisDay",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Date in YYYY-MM-DD format. Defaults to today if not provided.",
            "schema": {
              "example": "2026-02-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "On This Day memories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnThisDayResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get \"On This Day\" memories",
        "tags": [
          "Memories"
        ]
      }
    },
    "/api/v1/memories/context": {
      "get": {
        "description": "Returns memories relative to the selected date. Day mode returns memories from 1 month, 6 months, and 1 year ago. Week mode returns weekly summaries for the same intervals.",
        "operationId": "MemoriesController_getContext",
        "parameters": [
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "description": "Memory context mode",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week"
              ]
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Selected date in YYYY-MM-DD (ISO 8601)",
            "schema": {
              "example": "2026-02-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Context-based memories (day or week mode)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DayContextResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/WeekContextResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid date or mode"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get context-based memories for a selected date",
        "tags": [
          "Memories"
        ]
      }
    },
    "/api/v1/recommendations": {
      "get": {
        "operationId": "RecommendationsController_getRecommendations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns recommendation lists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get recommended categories and moods",
        "tags": [
          "Recommendations"
        ]
      }
    },
...
  },
  "info": {
    "title": "LifeSpan API",
    "description": "Backend API for Life Timeline — build a visual timeline of your life",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEventGroupDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "categoryId",
          "title"
        ]
      },
      "EventGroupCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ]
      },
      "EventPeriodResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "example": "2024-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2024-01-31",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "startDate",
          "createdAt"
        ]
      },
      "EventGroupResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventGroupCategoryDto"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPeriodResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "periods",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateEventGroupDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateEventPeriodDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "startDate"
        ]
      },
      "MoodStatDto": {
        "type": "object",
        "properties": {
          "dayStateName": {
            "type": "string"
          },
          "dayStateColor": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          }
        },
        "required": [
          "dayStateName",
          "dayStateColor",
          "count",
          "percentage"
        ]
      },
      "EventGroupDetailsMediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "s3Key",
          "url",
          "fileName",
          "contentType",
          "size",
          "createdAt"
        ]
      },
...
      "ChapterAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalPeriods": {
            "type": "number"
          },
          "totalDays": {
            "type": "number"
          },
          "totalMedia": {
            "type": "number"
          },
          "averageMoodScore": {
            "type": "number",
            "nullable": true
          },
          "moodDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoodDistributionItemDto"
            }
          },
          "density": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodDensityDto"
            }
          }
        },
        "required": [
          "totalPeriods",
          "totalDays",
          "totalMedia",
          "moodDistribution",
          "density"
        ]
      },
      "EventGroupDetailsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventGroupCategoryDto"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPeriodResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "moodStats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoodStatDto"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventGroupDetailsMediaDto"
            }
          },
          "totalDays": {
            "type": "number"
          },
          "analytics": {
            "$ref": "#/components/schemas/ChapterAnalyticsDto"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "periods",
          "createdAt",
          "updatedAt",
          "moodStats",
          "media",
          "totalDays",
          "analytics"
        ]
      },
...
      "DayMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "Presigned URL for reading the media file (null if generation failed)",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "s3Key",
          "url",
          "fileName",
          "contentType",
          "size",
          "createdAt"
        ]
      },
...
      "CreateDayMediaDto": {
        "type": "object",
        "properties": {
          "s3Key": {
            "type": "string",
            "example": "uploads/user-id/uuid.jpg",
            "description": "S3 object key"
          },
          "fileName": {
            "type": "string",
            "example": "photo.jpg",
            "description": "Original file name"
          },
          "contentType": {
            "type": "string",
            "example": "image/jpeg",
            "description": "MIME type",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/gif",
              "image/heic",
              "image/heif",
              "video/mp4",
              "video/webm",
              "video/quicktime"
            ]
          },
          "size": {
            "type": "number",
            "example": 1024000,
            "description": "File size in bytes"
          }
        },
        "required": [
          "s3Key",
          "fileName",
          "contentType",
          "size"
        ]
      },
...
      "DayMemoryDto": {
        "type": "object",
        "properties": {
          "interval": {
            "$ref": "#/components/schemas/IntervalDto"
          },
          "date": {
            "type": "string",
            "example": "2025-02-12"
          },
          "mood": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MoodDto"
              }
            ]
          },
          "mediaCount": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "interval",
          "date",
          "mediaCount"
        ]
      },
      "OnThisDayResponseDto": {
        "type": "object",
        "properties": {
          "baseDate": {
            "type": "string",
            "example": "2026-02-12"
          },
          "memories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayMemoryDto"
            }
          }
        },
        "required": [
          "baseDate",
          "memories"
        ]
      },
      "DayContextResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "day"
          },
          "baseDate": {
            "type": "string",
            "example": "2026-02-12"
          },
          "memories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayMemoryDto"
            }
          }
        },
        "required": [
          "type",
          "baseDate",
          "memories"
        ]
      },
      "BaseWeekDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "example": "2026-02-09"
          },
          "end": {
            "type": "string",
            "example": "2026-02-15"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
...
      "WeekContextResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "week"
          },
          "baseWeek": {
            "$ref": "#/components/schemas/BaseWeekDto"
          },
          "memories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeekMemoryDto"
            }
          }
        },
        "required": [
          "type",
          "baseWeek",
          "memories"
        ]
      },
      "RecommendationDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "work"
          },
          "color": {
            "type": "string",
            "example": "#3B82F6"
          }
        },
        "required": [
          "key",
          "color"
        ]
      },
      "RecommendationsResponseDto": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationDto"
            }
          },
          "moods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationDto"
            }
          }
        },
        "required": [
          "categories",
          "moods"
        ]
      },
...
    }
  }
}